requirements.txt
streamlitstreamlit
yfinance
pandas
plotly
streamlit-autorefresh

